+++ goods/repository.py
@@ -3,6 +3,7 @@ from graphql import GraphQLResolveInfo
 
 from app.errors import UnauthorizedError
 from app.repository_base import RepositoryBase, IRepository
+from category.models import Category
 from goods.models import Good
 from users.models import ExtendedUser
 
@@ -40,11 +41,32 @@ class GoodRepository(RepositoryBase, IRepository):
             get_by_id_base(GoodRepository, searched_id)
 
     @staticmethod
-    def update_item(item_id, **kwargs) -> [QuerySet]:
-        return super(GoodRepository, GoodRepository).\
-            update_item_base(GoodRepository, item_id=item_id, **kwargs)
+    def update_item(info: GraphQLResolveInfo, item_id, **kwargs) -> [QuerySet]:
+        user: ExtendedUser = info.context.user or None
+        good = Good.objects.filter(id=item_id).first()
+        if user.is_admin() or user == good.seller:
+            return super(GoodRepository, GoodRepository).\
+                update_item_base(GoodRepository, item_id=item_id, **kwargs)
+        else:
+            raise UnauthorizedError(
+                "Not enough permissions to call this endpoint")
 
     @staticmethod
-    def delete_item(searched_id: str):
-        return super(GoodRepository, GoodRepository). \
-            delete_item_base(GoodRepository, searched_id)
+    def delete_item(info: GraphQLResolveInfo, searched_id: str):
+        user: ExtendedUser = info.context.user or None
+        good = Good.objects.filter(id=searched_id).first()
+        if user.is_admin() or user.id == good.seller_id:
+            super(GoodRepository, GoodRepository). \
+                delete_item_base(GoodRepository, searched_id)
+
+    @staticmethod
+    def change_category(info: GraphQLResolveInfo,
+                        searched_id: str,
+                        category_id: str) -> [QuerySet]:
+        category = Category.objects.filter(id=category_id).first()
+        good = Good.objects.filter(id=searched_id).first()
+        user: ExtendedUser = info.context.user or None
+        if user.is_admin() or user == good.seller:
+            good.category = category
+            good.save()
+        return good
